mainをdiscordパッケージに分割
